refactor: 페이지 파일들을 종류별로 디렉토리 분리
